+++ NEWS.md
@@ -1,6 +1,6 @@
 # pmd 0.1.3
 
-- speed up GlobalStd by BiocParallel package
+- speed up GlobalStd by mapply
 - add message for getrda
 - add hmdb pmd analysis results as dataset